requirements.txt
asgiref
attrs
autobahn
Automat
cffi
channels
constantly
cryptography
daphne
Django==3.0.5
django-cors-headers
djangorestframework
djongo==1.3.3
dnspython
hyperlink
idna
incremental
pyasn1
pyasn1-modules
pycparser
PyHamcrest
pymongo
pyOpenSSL==19.1.0
python-dateutil
pytz
requests
service-identity
six
sqlparse==0.2.4
txaio
zope.interface
firebase-admin
jwcrypto